Move flat typemap property coverage to schema utils
