+++ test/logflare/source/bigquery/schema_test.exs
@@ -1,11 +1,9 @@
 defmodule Logflare.Sources.Source.BigQuery.SchemaTest do
   @moduledoc false
   use Logflare.DataCase
-  use ExUnitProperties
 
   alias GoogleApi.BigQuery.V2.Api.Tables, as: BigQueryTables
   alias Logflare.Google.BigQuery.SchemaUtils
-  alias Logflare.SourceSchemas
   alias Logflare.Sources.Source.BigQuery.Schema
 
   setup do
@@ -74,56 +72,6 @@ defmodule Logflare.Sources.Source.BigQuery.SchemaTest do
     # subsequent updates do not increase mock count
     le = build(:log_event, source: source, metadata: %{"change" => 123})
     assert :ok = Schema.update(pid, le, source)
-  end
-
-  property "generated metadata schemas include every field in the flat map" do
-    test_pid = self()
-
-    BigQueryTables
-    |> stub(:bigquery_tables_patch, fn _conn, _project_id, _dataset_id, _table_name, _opts ->
-      {:ok, %{}}
-    end)
-
-    Logflare.Mailer
-    |> stub(:deliver, fn _ ->
-      send(test_pid, :generated_schema_update_notification_delivered)
-      :ok
-    end)
-
-    check all(metadata <- metadata_generator(), max_runs: 50) do
-      user = insert(:user)
-      source = insert(:source, user: user)
-      schema = TestUtils.default_bq_schema()
-
-      insert(:source_schema,
-        source: source,
-        source_id: source.id,
-        bigquery_schema: schema,
-        schema_flat_map: SchemaUtils.bq_schema_to_flat_typemap(schema)
-      )
-
-      pid =
-        start_supervised!(
-          {Schema,
-           [
-             source: source,
-             plan: %{limit_source_fields_limit: 500},
-             bigquery_project_id: "some-id",
-             bigquery_dataset_id: "some-id"
-           ]},
-          id: {:schema, source.id}
-        )
-
-      le = build(:log_event, source: source, metadata: metadata)
-      assert :ok = Schema.update(pid, le, source)
-      assert_receive :generated_schema_update_notification_delivered, to_timeout(second: 30)
-
-      source_schema = SourceSchemas.get_source_schema_by(source_id: source.id)
-      expected_flat_map = expected_flat_map(metadata, "metadata")
-
-      assert Map.take(source_schema.schema_flat_map, Map.keys(expected_flat_map)) ==
-               expected_flat_map
-    end
   end
 
   test "does not patch when the incoming event matches the existing schema" do
@@ -197,96 +145,4 @@ defmodule Logflare.Sources.Source.BigQuery.SchemaTest do
     new_schema = TestUtils.build_bq_schema(%{"test" => "value"})
     Schema.notify_maybe(source.token, new_schema, old_schema)
   end
-
-  defp metadata_generator(depth \\ 2) do
-    StreamData.map_of(key_generator(), metadata_value_generator(depth),
-      min_length: 1,
-      max_length: 4
-    )
-  end
-
-  defp metadata_value_generator(0) do
-    StreamData.one_of([
-      scalar_generator(),
-      scalar_list_generator()
-    ])
-  end
-
-  defp metadata_value_generator(depth) do
-    StreamData.one_of([
-      scalar_generator(),
-      scalar_list_generator(),
-      metadata_generator(depth - 1),
-      StreamData.list_of(metadata_generator(depth - 1), min_length: 1, max_length: 1)
-    ])
-  end
-
-  defp key_generator do
-    first_chars = string_chars(?A..?Z) ++ string_chars(?a..?z) ++ ["_"]
-    chars = first_chars ++ string_chars(?0..?9)
-
-    StreamData.map(
-      {StreamData.member_of(first_chars),
-       StreamData.list_of(StreamData.member_of(chars), max_length: 7)},
-      fn {first, rest} -> first <> Enum.join(rest) end
-    )
-  end
-
-  defp string_chars(range), do: Enum.map(range, &<<&1::utf8>>)
-
-  defp scalar_generator do
-    StreamData.one_of([
-      StreamData.string(:alphanumeric, min_length: 1, max_length: 12),
-      StreamData.integer(),
-      StreamData.float(min: -1_000.0, max: 1_000.0),
-      StreamData.boolean()
-    ])
-  end
-
-  defp scalar_list_generator do
-    StreamData.one_of([
-      StreamData.list_of(StreamData.string(:alphanumeric, min_length: 1, max_length: 12),
-        min_length: 1,
-        max_length: 4
-      ),
-      StreamData.list_of(StreamData.integer(), min_length: 1, max_length: 4),
-      StreamData.list_of(StreamData.float(min: -1_000.0, max: 1_000.0),
-        min_length: 1,
-        max_length: 4
-      ),
-      StreamData.list_of(StreamData.boolean(), min_length: 1, max_length: 4)
-    ])
-  end
-
-  defp expected_flat_map(map, prefix) do
-    Enum.reduce(map, %{prefix => :map}, fn {key, value}, acc ->
-      path = prefix <> "." <> key
-
-      Map.merge(acc, expected_flat_entry(value, path))
-    end)
-  end
-
-  defp expected_flat_entry(value, path) when is_map(value), do: expected_flat_map(value, path)
-
-  defp expected_flat_entry(value, path) when is_list(value) do
-    value
-    |> Enum.filter(&is_map/1)
-    |> expected_flat_entry_from_list(value, path)
-  end
-
-  defp expected_flat_entry(value, path), do: %{path => scalar_type(value)}
-
-  defp expected_flat_entry_from_list([], value, path),
-    do: %{path => {:list, scalar_type(hd(value))}}
-
-  defp expected_flat_entry_from_list(maps, _value, path) do
-    Enum.reduce(maps, %{path => :map}, fn item, acc ->
-      Map.merge(acc, expected_flat_map(item, path))
-    end)
-  end
-
-  defp scalar_type(value) when is_binary(value), do: :string
-  defp scalar_type(value) when is_integer(value), do: :integer
-  defp scalar_type(value) when is_float(value), do: :float
-  defp scalar_type(value) when is_boolean(value), do: :boolean
 end
